Fix countries issue

+++ server/api/user/[id].patch.ts
@@ -67,30 +67,26 @@ export default defineEventHandler(async (event) => {
       where: { id },
       data: filteredData,
     });
-
-    if (user.image[0]) {
-      await deleteFromCloudinary(user.image[0].public_id, "image");
+    if (user) {
+      if (user.image[0]) {
+        await deleteFromCloudinary(user.image[0].public_id, "image");
+      }
+      if (user.qr_code[0]) {
+        await deleteFromCloudinary(user.qr_code[0].public_id, "image");
+      }
+      if (user.pdf_file[0]) {
+        await deleteFromCloudinary(user.pdf_file[0].public_id, "file");
+      }
+      await prisma.pdfFile.deleteMany({
+        where: { user_id: user.id },
+      });
+      await prisma.image.deleteMany({
+        where: { user_id: user.id },
+      });
+      await prisma.qrCode.deleteMany({
+        where: { user_id: user.id },
+      });
     }
-    if (user.qr_code[0]) {
-      await deleteFromCloudinary(user.qr_code[0].public_id, "image");
-    }
-    if (user.pdf_file[0]) {
-      await deleteFromCloudinary(user.pdf_file[0].public_id, "file");
-    }
-    await prisma.pdfFile.deleteMany({
-      where: { user_id: user.id },
-    });
-    await prisma.image.deleteMany({
-      where: { user_id: user.id },
-    });
-    await prisma.qrCode.deleteMany({
-      where: { user_id: user.id },
-    });
-
-    await prisma.user.delete({
-      where: { id },
-    });
-
     return {
       message: "User updated successfully.",
       result,
